+++ Source/RocketSoundEnhancement/RocketSoundEnhancement.cs
@@ -209,15 +209,14 @@ namespace RocketSoundEnhancement
                 {
                     source.outputAudioMixerGroup = sourcePart?.vessel == FlightGlobals.ActiveVessel ? FocusMixer : ExteriorMixer;
 
-                    if (Settings.MufflerQuality > AudioMufflerQuality.Normal)
+                    if (Settings.MufflerQuality > AudioMufflerQuality.Normal && Settings.MachEffectsAmount > 0)
                     {
+                        float machPass = Mathf.Lerp(Settings.MachEffectLowerLimit, 1, sourcePart.vessel.GetComponent<ShipEffects>().MachPass);
+
                         if (!managedMinDistance.ContainsKey(managedSourceID))
                             managedMinDistance.Add(managedSourceID, source.minDistance);
 
-                        float machPass = Mathf.Lerp(Settings.MachEffectLowerLimit, 1, sourcePart.vessel.GetComponent<ShipEffects>().MachPass);
-                        float sourceDistance = Vector3.Distance(CameraManager.GetCurrentCamera().transform.position, source.transform.position);
-                        float distanceAttenuation = Mathf.Max(Mathf.Pow(1 - Mathf.Clamp01(sourceDistance / Settings.AirSimMaxDistance), 10), 0.1f) * machPass;
-                        source.minDistance = managedMinDistance[managedSourceID] * distanceAttenuation;
+                        source.minDistance = managedMinDistance[managedSourceID] * machPass;
                         continue;
                     }
 
@@ -226,14 +225,11 @@ namespace RocketSoundEnhancement
                         source.minDistance = managedMinDistance[managedSourceID];
                         managedMinDistance.Remove(managedSourceID);
                     }
-
-                    if (source.GetComponent<AirSimulationFilter>()) UnityEngine.Object.Destroy(source.GetComponent<AirSimulationFilter>());
-
                     continue;
                 }
 
                 source.outputAudioMixerGroup = ExteriorMixer;
-                if (Settings.MufflerQuality == AudioMufflerQuality.AirSim && source.name.StartsWith("Explosion"))
+                if (Settings.MufflerQuality == AudioMufflerQuality.AirSim && source.gameObject.GetComponents<AudioSource>().Length == 1)
                 {
                     var airSimFilter = source.gameObject.AddOrGetComponent<AirSimulationFilter>();
                     airSimFilter.enabled = true;